performance-analysis/requirements.txt



fastapi==0.111.1
fastapi-cli==0.0.4


hyperopt==0.2.7



matplotlib-inline==0.1.7

numpy==1.26.4

optuna==4.2.0


pandas==2.2.3






requests==2.32.3
requests-oauthlib==1.3.0

scikit-learn==1.5.0
shap==0.46.0

xgboost==2.1.4
seaborn

